refactor: split benchmark reference ingestion dtos
